mango leaf gradio/requirements.txt
gradio
tensorflow
pillow
numpy